+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 02/07/2026
+
+## 1.2.0
+
+- Updating the UX of encoding progress
+  - Now displays the encoding percentage
+  - Now displays an approximate ETA until encoding is finished
+  - Now updates a single line instead of printing one line per progress step
+- Iterations are now timed and the time an iteration took is displayed after encoding is done
+- Total time taking for the command is now displayed in the final results
+
 # 02/05/2026
 
 ## 1.1.0